requirements.txt
greenlet==3.4.0
python-jose[cryptography]==3.5.0
PySide6==6.11.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
shiboken6==6.11.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
